+++ backend/src/main/java/com/Swoop/Swoop/controllers/SwoopDao.java
@@ -14,7 +14,7 @@ public interface SwoopDao {
     Optional<User> selectUserByEmail(String email);
     int deleteUserByEmail(String email);
     int updateUserByEmail(String email, User user);
-	void addRide(User user, String start, String end, int distance);
+	void addRide(User user, String start, String end, float distance);
 	Map<Integer, Ride> getRides();
 	void finishRide(User u);
 	Ride getRideByID(int currentRideID);
